requirements.txt
flask
firebase-admin
google-cloud-firestore
joblib
librosa
numpy
requests
